meta: update tests from #14505 (#15120)
* meta: update unit tests from #14505

* meta: update typing tests

* meta: update integration tests from #14505

* meta: fix ibmi/db2 unit test

* meta: fix failing integration tests

* meta: undo integration/utils changes

* meta: undo index order

* meta: remove incorrect mention of bigint

* meta: fix index order test

+++ test/integration/model/notExist.test.js
@@ -11,7 +11,7 @@ describe(Support.getTestDialectTeaser('Model'), () => {
     this.Order = this.sequelize.define('Order', {
       id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
       sequence: DataTypes.INTEGER,
-      amount: DataTypes.DECIMAL,
+      amount: DataTypes.FLOAT,
       type: DataTypes.STRING,
     });
